requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Werkzeug==3.0.3
APScheduler==3.10.4
python-dateutil==2.9.0
email-validator==2.2.0
